[Refactor] #124 - 액션시트 코드 수정
- 생성자에서 주입되어야 하는 내용 간결화 (타이틀과 메세지만 주입)
- UIAlertController 처럼 addAction하여 액션 및 버튼을 추가하는 형태로 구현
